+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for customizing options
+
 ### Updated
 
 - Copy views postinstall
 - About GTFS-to-HTML
 - Dependency updates
-
 
 ## [1.2.0] - 2024-09-08